refactor data import in chart components and add data.json file
