+++ frontend/pages/home/data.json
@@ -1,0 +1,57 @@
+[
+  {
+    "sdk_key": "550e8400-e29b-41d4-a716-446655440000",
+    "request_id": "123e4567-e89b-12d3-a456-426614174000",
+    "raw_req_id": "REQ001-001",
+    "raw_resp_id": "RESP001-001",
+    "req_in": "2024-11-16T12:30:45Z",
+    "req_out": "2024-11-16T12:30:50Z",
+    "endpoint": "/api/v1/resource",
+    "resp_states": "200 OK",
+    "method": "GET"
+  },
+  {
+    "sdk_key": "d3b07384-d9a2-4b5d-96b9-d44d30821c17",
+    "request_id": "223e4567-e89b-12d3-a456-426614174001",
+    "raw_req_id": "REQ002-002",
+    "raw_resp_id": "RESP002-002",
+    "req_in": "2024-11-16T13:00:00Z",
+    "req_out": "2024-11-16T13:00:05Z",
+    "endpoint": "/api/v1/login",
+    "resp_states": "401 Unauthorized",
+    "method": "POST"
+  },
+  {
+    "sdk_key": "a8d4b167-cbf1-403b-b1f3-5c517b8ef2b0",
+    "request_id": "323e4567-e89b-12d3-a456-426614174002",
+    "raw_req_id": "REQ003-003",
+    "raw_resp_id": "RESP003-003",
+    "req_in": "2024-11-16T13:30:15Z",
+    "req_out": "2024-11-16T13:30:20Z",
+    "endpoint": "/api/v1/logout",
+    "resp_states": "200 OK",
+    "method": "DELETE"
+  },
+  {
+    "sdk_key": "f47ac10b-58cc-4372-a567-0e02b2c3d479",
+    "request_id": "423e4567-e89b-12d3-a456-426614174003",
+    "raw_req_id": "REQ004-004",
+    "raw_resp_id": "RESP004-004",
+    "req_in": "2024-11-16T14:00:30Z",
+    "req_out": "2024-11-16T14:00:40Z",
+    "endpoint": "/api/v1/data",
+    "resp_states": "500 Internal Server Error",
+    "method": "GET"
+  },
+  {
+    "sdk_key": "c9f0f895-dedf-4d49-b045-9a2b0e1e5874",
+    "request_id": "523e4567-e89b-12d3-a456-426614174004",
+    "raw_req_id": "REQ005-005",
+    "raw_resp_id": "RESP005-005",
+    "req_in": "2024-11-16T14:00:30Z",
+    "req_out": "2024-11-16T14:00:40Z",
+    "endpoint": "/api/v1/upload",
+    "resp_states": "201 Created",
+    "method": "POST"
+  }
+]
